move shallow to @tanstack/store
